Run on all available devices
